requirements.txt
# YouTube Automation Dependencies
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
ffmpeg-python
requests
python-dotenv
Pillow
firebase-admin
aiohttp
pandas
APScheduler
flask
opencv-python
numpy
